feature: removed 'sam init' for addComponents
