+++ bin/index.ts
@@ -4,7 +4,7 @@ import * as modules from "../src/resources/modules/modules";
 import * as components from "../src/resources/components/components";
 import * as generateSAM from "../src/generateSAM/generatesam";
 import * as generateCustomSAM from "../src/generateSAM/customSAM";
-import * as addComponents from "../src/addComponents/addcomponents";
+import * as addComponent from "../src/addComponents/addComponent";
 import * as addModules from "../src/addModules/addModules";
 import * as addModulesToExisting from "../src/addModulesToexisting/addModulesToExisting";
 import * as generateSAMTypes from "../src/generateSAM/generatesam.types";
@@ -16,7 +16,7 @@ export = {
   generateSAM: generateSAM,
   generateSAMTypes: generateSAMTypes,
   generateCustomSAM: generateCustomSAM,
-  addComponents: addComponents,
+  addComponent: addComponent,
   addModules: addModules,
   addModulesToExisting: addModulesToExisting,
 };
