+++ src/generateSAM/customSAM.ts
@@ -4,7 +4,7 @@ import {
   OpenAIApi,
 } from "openai";
 import * as dotenv from "dotenv"; // see https://github.com/motdotla/dotenv#how-do-i-use-dotenv-with-import
-import { copyRecursiveSync } from "../addModulesToexisting/addModulesToExisting";
+import { copyRecursiveSync } from "../utlities/generate-utilities";
 import * as utlities from "../utlities/utilities";
 import * as Yaml from "js-yaml";
 import * as fs from "fs";
